requirements.txt
Flask==2.1.0
convertapi==1.7.0
python-pptx==0.6.23
werkzeug==2.1.0
openai==1.13.3
google-generativeai==0.4.0
moviepy==1.0.3
boto3==1.18.65


